SDK-1659; Refactor code into separate sub-packages
Create: web, shareurl, util, profile. aml
