+++ go.mod
@@ -2,7 +2,6 @@ module github.com/getyoti/yoti-go-sdk/v3
 
 require (
 	github.com/golang/protobuf v1.4.2
-	github.com/pkg/errors v0.9.1 // indirect
 	gotest.tools/v3 v3.0.2
 )
