--- http_client.go
@@ -1,10 +1,0 @@
-package yoti
-
-import (
-	"net/http"
-)
-
-//HttpClient is a mockable HTTP Client Interface
-type HttpClient interface {
-	Do(*http.Request) (*http.Response, error)
-}
