+++ profile/sandbox/client_test.go
@@ -95,7 +95,7 @@ func TestClient_SetupSharingProfileUsesDefaultUrlAsFallbackWithNoEnvValue(t *tes
 }
 
 func createSandboxClient(t *testing.T, constructorBaseUrl string) (client Client) {
-	keyBytes, fileErr := ioutil.ReadFile("../../test-key.pem")
+	keyBytes, fileErr := ioutil.ReadFile("../../test/test-key.pem")
 	assert.NilError(t, fileErr)
 
 	pemFile, parseErr := cryptoutil.ParseRSAKey(keyBytes)
